refactor(trie): hashed post state
