+++ crates/blockchain-tree/src/blockchain_tree.rs
@@ -1174,12 +1174,9 @@ impl<DB: Database, EF: ExecutorFactory> BlockchainTree<DB, EF> {
             }
             None => {
                 debug!(target: "blockchain_tree", blocks = ?block_hash_numbers, "Recomputing state root for insert");
-                let (state_root, trie_updates) = state
-                    .state_root_calculator(
-                        self.externals.provider_factory.provider()?.tx_ref(),
-                        &hashed_state,
-                    )
-                    .root_with_updates()
+                let provider = self.externals.provider_factory.provider()?;
+                let (state_root, trie_updates) = hashed_state
+                    .state_root_with_updates(provider.tx_ref())
                     .map_err(Into::<DatabaseError>::into)?;
                 let tip = blocks.tip();
                 if state_root != tip.state_root {
