+++ crates/trie/Cargo.toml
@@ -16,9 +16,12 @@ workspace = true
 reth-primitives.workspace = true
 reth-interfaces.workspace = true
 reth-db.workspace = true
+revm.workspace = true
 
+# alloy
 alloy-rlp.workspace = true
 alloy-chains.workspace = true
+
 # tokio
 tokio = { workspace = true, default-features = false, features = ["sync"] }
